+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 * Bump compileSdkVersion to apiLevel 30
   [#1202](https://github.com/bugsnag/bugsnag-android/pull/1202)
 
+* Collect whether the system has restricted background work for the app
+  [#1211](https://github.com/bugsnag/bugsnag-android/pull/1211)
+
 * Add public API for crash-on-launch detection
   [#1157](https://github.com/bugsnag/bugsnag-android/pull/1157)
   [#1159](https://github.com/bugsnag/bugsnag-android/pull/1159)